Adicionando Parte de Professores

+++ src/php/Auxiliares/alterAvatar.php
@@ -7,13 +7,10 @@
     }
     
     if (isset($_SESSION['matricula'])) {
+        
         $matricula = $_SESSION['matricula'];
-        if (substr($matricula,0,1) == 0) {
-            $dadosUsuario = $sql -> query("SELECT avatar FROM aluno WHERE matricula = '$matricula'");
-        }
-        else{
-            $dadosUsuario = $sql -> query("SELECT avatar FROM professor WHERE matricula = '$matricula'");
-        }
+        $usuario = substr($matricula,0,1) == 0 ? 'aluno' : 'professor';
+        $dadosUsuario = $sql -> query("SELECT avatar FROM $usuario WHERE matricula = '$matricula'");
     }
     else{
         header('Location: ../cadastro_login.html');
@@ -27,22 +24,14 @@
     while ($dados = mysqli_fetch_array($dadosUsuario)) {
         $avatarA = $dados['avatar'];
     }
+    
+    if ($avatarA != 'd_img.png') { unlink('../../assets/img/Avatares/' . $avatarA); }
 
-    if ($avatarA == 'd_img.png') {
-        move_uploaded_file($_FILES['avatar']['tmp_name'],$avatar);
-        $sql -> query(
-            "UPDATE aluno SET
-                `avatar` = '$avatar'
-            WHERE matricula = '$matricula'");
-    }
-    else {
-        unlink('../../assets/img/Avatares/' . $avatarA);
-        move_uploaded_file($_FILES['avatar']['tmp_name'],'../../assets/img/Avatares/'.$avatar);
-        $sql -> query(
-            "UPDATE aluno SET
-                `avatar` = '$avatar'
-            WHERE matricula = '$matricula'");
-    }
+    move_uploaded_file($_FILES['avatar']['tmp_name'],'../../assets/img/Avatares/'.$avatar);
+    $sql -> query(
+        "UPDATE $usuario SET
+            `avatar` = '$avatar'
+        WHERE matricula = '$matricula'");
     
     header("Location: ../perfil.php");
